// src/types/campaign.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface Campaign {
  name: string;
  description?: string;
  image?: {
    /**
     * data URL of embedded image
     */
    url: string;
    /**
     * alt text for the image
     */
    alt?: string;
    [k: string]: unknown;
  };
  link?: {
    /**
     * URL to link to
     */
    url: string;
    /**
     * text to display for the link
     */
    text?: string;
    [k: string]: unknown;
  };
  funding?: {
    /**
     * symbol of a token that will be given away as part of the campaign
     */
    token: string;
    /**
     * amount of the token that will be given away as part of the campaign
     */
    amount: number;
    [k: string]: unknown;
  }[];
  /**
   * @minItems 1
   * @maxItems 5
   */
  tags:
    | [string]
    | [string, string]
    | [string, string, string]
    | [string, string, string, string]
    | [string, string, string, string, string];
  [k: string]: unknown;
}